[Tools][GTK][browserperfdash-benchmark] browser-binary-size plan plugin fails to find the browser libraries in the GTK performance bot https://bugs.webkit.org/show_bug.cgi?id=303174
Reviewed by Nikolas Zimmermann and Carlos Garcia Campos.

If the DT_NEEDED entries for WebKit libraries in the MiniBrowser binary contain full paths,
and those paths point to symlinks outside the build directory that ultimately resolve to
locations inside the build directory, then the browser-binary-size plugin incorrectly
classifies those libraries as external when they should be considered part of the build.

This is a corner case that occurs when you build WebKit in one directory, then move the
WebKit checkout to another location and create a symlink from the original directory to
the new one. In this scenario, the LD_LIBRARY_PATH environment variable set by the test
cannot override DT_NEEDED entries with full paths. In practice, this isn't a functional
issue because the correct library still gets loaded, but it causes the plugin to
incorrectly identify the library as being outside the build directory.

Fix this by also checking whether the resolved path belongs to the build directory.

* Tools/Scripts/webkitpy/browserperfdash/plans/browser_binary_size.py:
(get_browser_relevant_objects_glib):

Canonical link: https://commits.webkit.org/303603@main