[GStreamer][WebRTC] Move network access to NetworkProcess https://bugs.webkit.org/show_bug.cgi?id=244002
Reviewed by Xabier Rodriguez-Calvar and Adrian Perez de Castro.

This patch introduces a build-time dependency on librice for the GTK port. The USE_LIBRICE CMake
option can be set to FALSE if this dependency is not desired. It is FALSE by default on WPE.

When USE_LIBRICE is enabled, the WebRTC network handling is now deferred to the NetworkProcess,
bi-directional communication is performed with the WebProcess which makes use of a custom
GstWebRTCICE implementation. Runtime to the previous un-sandboxed libnice-based implementation can
be done by setting the WEBKIT_GST_DISABLE_WEBRTC_NETWORK_SANDBOX=1 environment variable.

Canonical link: https://commits.webkit.org/303623@main