[threaded-animations] WPT test scroll-animations/css/animation-timeline-view-functional-notation.tentative.html crashes with "Threaded Scroll-driven Animations" enabled https://bugs.webkit.org/show_bug.cgi?id=303254 rdar://165550962
Reviewed by Tim Nguyen.

Check that we do indeed have an animation before calling `acceleratedStateDidChange()` on it.

Note that there is no test change in this patch since the flag is not yet enabled on bots. This
was caught in preparation of that running animation tests locally using
`--experimental-feature ThreadedScrollDrivenAnimationsEnabled=true`.

* Source/WebCore/animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::addPendingAcceleratedAction):

Canonical link: https://commits.webkit.org/303646@main