[threaded-animations] WPT test `scroll-animations/css/animation-timeline-scroll-functional-notation.tentative.html` fails with "Threaded Scroll-driven Animations" enabled https://bugs.webkit.org/show_bug.cgi?id=303264 rdar://165562929
Reviewed by Anne van Kesteren.

This test would fail because we failed to account for the `UseCachedCurrentTime` parameter passed to
`ScrollTimeline::currentTime()`. And if we query the timeline current time for an accelerated animation,
we pass the value `UseCachedCurrentTime::No`.

We now move the current time data computation logic into a dedicated method, call it when caching the
current time, and calling it as well if `UseCachedCurrentTime::No` is passed when querying the current
time.

Note that there is no test change in this patch since the flag is not yet enabled on bots. This
was caught in preparation of that running animation tests locally using
`--experimental-feature ThreadedScrollDrivenAnimationsEnabled=true`.

* Source/WebCore/animation/ScrollTimeline.cpp:
(WebCore::ScrollTimeline::computeCurrentTimeData const):
(WebCore::ScrollTimeline::cacheCurrentTime):
(WebCore::ScrollTimeline::computeTimelineData const):
(WebCore::ScrollTimeline::currentTime):
* Source/WebCore/animation/ScrollTimeline.h:
* Source/WebCore/animation/ViewTimeline.cpp:
(WebCore::ViewTimeline::computeTimelineData const):
* Source/WebCore/animation/ViewTimeline.h:

Canonical link: https://commits.webkit.org/303650@main