[GStreamer][WebRTC] Add support for framesPerSecond and totalDecodeTime for remote inbound RTP stats https://bugs.webkit.org/show_bug.cgi?id=303261
Reviewed by Xabier Rodriguez-Calvar.

The frames-per-second value is derived from the video sink avg-rate statistics applied to the
framerate negotiated with the caps. The totalVideoDecodeTime is the the accumulated time used for
video decoding, we already measured this as part of the video frame metadata gathering. The stats
reported by the incoming track processor are used when it is in charge of video decoding. Otherwise
the stats are collected at the media player level.

Canonical link: https://commits.webkit.org/303671@main