Add an event that fires for user text input https://bugs.webkit.org/show_bug.cgi?id=302969 rdar://problem/165226356
Reviewed by Per Arne Vollan.

Some use cases want an event that only fires when real user typing occurs (i.e. not synthetic
modifications to input or textarea via JS). Adding an event handler for the input event that checks
for isTrusted events mostly accomplishes this. But it still has some overhead that we don't want to
pay for (e.g. an untrusted click event on a checkbox causes a trusted input event to fire, which
then spuriously invokes the code that only cares about real user typing).

To help with this, add a webkitusertextinput event that only fires when real user typing occurs
which modifies a text field or text area. This is analogous to the behavior of the
textDidChangeInTextField and textDidChangeInTextArea APIs that we offer to the injected bundle.

For now, this event is only available to autofill worlds.

Test: Tools/TestWebKitAPI/Tests/mac/KeyboardEventTests.mm
* Source/WebCore/dom/EventNames.json:
* Source/WebCore/html/HTMLTextFormControlElement.cpp:
(WebCore::HTMLTextFormControlElement::dispatchUserTextInputEvent):
* Source/WebCore/html/HTMLTextFormControlElement.h:
* Source/WebKit/WebProcess/WebCoreSupport/WebEditorClient.cpp:
(WebKit::WebEditorClient::textDidChangeInTextField):
(WebKit::WebEditorClient::textDidChangeInTextArea):
* Tools/TestWebKitAPI/Tests/mac/KeyboardEventTests.mm:
(TestWebKitAPI::TEST(KeyboardEventTests, UserTextInputEvent)):

Canonical link: https://commits.webkit.org/303685@main