Text box makes text-decoration:underline higher than it should be https://bugs.webkit.org/show_bug.cgi?id=303082 <rdar://problem/165614136>
Reviewed by Antti Koivisto.

When text-box-trim is applied to the root inline box, we move the entire box with its
content. In such case we don't need to offset the underline unlike when the content moves
within the root inline box (when text-box-trim is on a non-root inline box).

Test: fast/inline/text-box-trim-with-underline3.html

* LayoutTests/fast/inline/text-box-trim-with-underline3-expected.html: Added.
* LayoutTests/fast/inline/text-box-trim-with-underline3.html: Added.
* LayoutTests/platform/ios/TestExpectations:
* Source/WebCore/style/InlineTextBoxStyle.cpp:
(WebCore::underlineOffsetForTextBoxPainting):

Canonical link: https://commits.webkit.org/303710@main