[GFC] Move gap computation to GridLayoutUtils https://bugs.webkit.org/show_bug.cgi?id=303351 rdar://165662046
Reviewed by Vitor Roriz and Tim Nguyen.

Currently only GridFormattingContext uses gaps as part of determining
the location of the grid items after layout, but soon GridLayout will
need to know the gaps for the columns and rows as well. So that both
classes can have access to this logic we can have it live as part of
GridLayoutUtils for now.

Canonical link: https://commits.webkit.org/303731@main