Add a basic implementation of ReadableStream async iterable rdar://165605714 https://bugs.webkit.org/show_bug.cgi?id=303302
Reviewed by Chris Dumez.

We add a basic ReadableStream Iterator implementation.
As per https://streams.spec.whatwg.org/#rs-asynciterator, we implement the init steps and the get next iteration result steps.
A follow-up will implement the return steps.

This is implemented in ReadableStream::Iterator class which takes a reader from its stream.
Given it might throw, we add support for throwing in iteratorCreate.

To do so, we create the internal iterator and pass it to JSDOMIteratorBase or JSDOMAsyncIteratorBase.
We add support for passing parameters to the iterator creation steps and make use of it for passing ReadableStreamIteratorOptions.

We do not yet support converting iterator creation parameters, the binding generator only creates default values (in this case preventCancel = false). This will be fixed in a follow-up.

We also update the binding generator to not expose the entries and keys methods for async iterables that are not key value.
We do not do so for sync iterator since that would break FontFaceSet (which should be moved to setlike).

We update the binding generator according the new iteratorCreate function.

Covered by rebased WPT tests and rebased binding tests.

Canonical link: https://commits.webkit.org/303734@main