OpaqueJSString::tryCreate(String&&) omits isolatedCopy https://bugs.webkit.org/show_bug.cgi?id=303299 rdar://165604488
Reviewed by Keith Miller.

As per the const String& constructor, OpaqueJSString intends to
hold an isolated copy of the string. Not all String&& are new strings,
so they need the isolatedCopy() call, too.

Remove const String& constructor, as that is redudundant when the
intention is to always adopt the parameter into a member. Use the
String&& constructor.

* Source/JavaScriptCore/API/OpaqueJSString.cpp:
(OpaqueJSString::tryCreate):
* Source/JavaScriptCore/API/OpaqueJSString.h:
(OpaqueJSString::OpaqueJSString):

Canonical link: https://commits.webkit.org/303743@main