Nest modules in WebKit internal modulemap https://bugs.webkit.org/show_bug.cgi?id=302948 rdar://165212021
Reviewed by Richard Robinson.

Previously, two top-level modules were given in the WebKit internal modulemap.
This appears to have caused clang some parsing confusion - the 'requires objc'
directive of the first module was taken to apply to the second. (This issue is
known as rdar://165045205 but is probably working as designed).

This commit adjusts the modulemap to nest them in a top-level module. Because
the submodules do not contain the 'explicit' keyword this also means we do not
need to spell out the full path to each module in users; the Swift demo logo
code is duly adjusted.

* Source/WebKit/Modules/Internal/module.modulemap:
* Source/WebKit/UIProcess/SwiftDemoLogo.swift:

Canonical link: https://commits.webkit.org/303747@main