The `initial-value` descriptor when registering Custom Properties can be optional https://bugs.webkit.org/show_bug.cgi?id=276223 rdar://131288198
Reviewed by Antti Koivisto.

This patch aligns WebKit with Web Specification [1]:

[1] https://drafts.css-houdini.org/css-properties-values-api/#initial-value-descriptor

"If the value of the syntax descriptor is the universal syntax definition, then
the initial-value descriptor is optional. If omitted, the initial value of the
property is the guaranteed-invalid value."

This patch removes early return in cases of parsing declaration value and initial
value for universal syntax.

Test: imported/w3c/web-platform-tests/css/css-properties-values-api/at-property-optional-initial-value.html
* LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/at-property-optional-initial-value-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/at-property-optional-initial-value.html: Added.
* Source/WebCore/css/parser/CSSVariableParser.cpp:
(WebCore::CSSVariableParser::parseDeclarationValue):
(WebCore::CSSVariableParser::parseInitialValueForUniversalSyntax):

Canonical link: https://commits.webkit.org/303757@main