REGRESSION(303317@main): [GStreamer] Crash in VideoFrame::copyTo() https://bugs.webkit.org/show_bug.cgi?id=303365
Reviewed by Xabier Rodriguez-Calvar.

The Y plane in I420 frames is not sub-sampled, so each row spans on the width of the frame. For
odd-sized frames we need to use the plane stride though (this is covered by existing WPT tests).
Driving-by, apply the same change for the A plane of I420A frames, which is not sub-sampled either.

* LayoutTests/platform/glib/TestExpectations:
* Source/WebCore/platform/graphics/gstreamer/VideoFrameGStreamer.cpp:
(WebCore::VideoFrame::copyTo):

Canonical link: https://commits.webkit.org/303818@main