[Filters] Add GraphicsContextFiltersEnabled preference back as a stable feature https://bugs.webkit.org/show_bug.cgi?id=303430 rdar://165725017
Reviewed by Simon Fraser.

This feature flag was removed in 294623@main. It was added back a mature feature
in 297006@main. This made it checkable in the code but invisible in the UI.

A separate feature flag was added later for GaussianBlur filter in 299740@main.
This feature flag is still stable so it allows turning the CG GaussianBlur from
the UI.

Some bugs were reported recently not related to GaussianBlur like bug 302900.
Having GraphicsContextFiltersEnabled visible in the UI would have made it easier
to confirm whether these bugs are related to CoreGraphics filter. Also having a
separate feature flag for GaussianBlur made the code a little bit complicated.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebCore/page/Page.cpp:
(WebCore::Page::preferredFilterRenderingModes const):
* Source/WebCore/platform/graphics/filters/FEGaussianBlur.cpp:
(WebCore::FEGaussianBlur::supportedFilterRenderingModes const):
* Source/WebCore/platform/graphics/filters/FilterRenderingMode.cpp:
(WebCore::operator<<):
* Source/WebCore/platform/graphics/filters/FilterRenderingMode.h:

Canonical link: https://commits.webkit.org/303819@main