ReadableByteStreamController::respondWithNewView is called from the tee algorithm with an empty queue rdar://165613840 https://bugs.webkit.org/show_bug.cgi?id=303380
Reviewed by Chris Dumez.

As part of running the close steps for byte tee, we call close on the stream branch1 and branch2.
Closing might throw an exception in case we have started to write data in an ongoing byob request.
We need to stop executing the close steps in that case, which will prevent the ReadableByteStreamController::respondWithNewView call.
We do this after closing both branches of the tee.

We also fix ReadableByteStreamController::shouldCallPull as a fly-by fix.

Test: streams/blob-and-tee.html
Canonical link: https://commits.webkit.org/303821@main