Add a basic implementation of ReadableStream async iterable return method rdar://165606483 https://bugs.webkit.org/show_bug.cgi?id=303303
Reviewed by Chris Dumez.

As per https://webidl.spec.whatwg.org/#js-asynchronous-iterator-prototype-object, we add support for the return method,
which allows to do error handling when iterating.

We add a concept to detect whether a native iterator implements returnSteps. If so, we expose the return method.
The returnSteps method returns a Ref<DOMPromise>.

We update JSDOMAsyncIterator to implement https://webidl.spec.whatwg.org/#js-asynchronous-iterator-prototype-object.
We reuse the settle method to handle both runNextSteps and runReturnSteps.
To disambiguate between the two, the settle method will be called with a parameter for runReturnSteps and without for runNextSteps.

Covered by rebased tests.

Canonical link: https://commits.webkit.org/303823@main