[GTK][WPE] Allow to schedule a composition for non rendering update requests while waiting for tiles https://bugs.webkit.org/show_bug.cgi?id=303379
Reviewed by Miguel Gomez.

We currently block any composition requests when we are waiting for
tiles. This patch allows to run the composition for other reasons while
waiting for the tiles of a request update that will be processed once the
tiles are ready. The state WaitingForTiles has been removed from the
enum and converted into a bool flag, since now we can be waiting for
tiles in any state. If rendering update is the only reason for the
composition we still wait for the tiles to be ready, switching to
Scheduled state but without activating the render timer. If another
composition request happens while waiting for tiles we keep in the
Scheduled state, but now activating the timer. When the timer fires, if
we are still waiting for the tiles, we remove all reasons but the
rendering update. Once current frame is completed we check the pending
reasons and if we are waiting for tiles we switch again to Scheduled
state without activating the timer.

Canonical link: https://commits.webkit.org/303824@main