[OpenXR] ASSERT when creating projection layers https://bugs.webkit.org/show_bug.cgi?id=303461
Reviewed by Fujii Hironori.

When creating a projection layer, the completion handler was running in
the thread of the run loop that handles the calls to OpenXR. However it
was created in the main thread, and thus it was asserting in Debug
build. By using callonMainRunLoop we ensure that the handler is called
in the right thread.

* Source/WebKit/UIProcess/XR/openxr/PlatformXROpenXR.cpp:
(WebKit::OpenXRCoordinator::createLayerProjection):

Canonical link: https://commits.webkit.org/303827@main