ReadableStream readers should not be GCed if having pending read requests for native sources rdar://165746977 https://bugs.webkit.org/show_bug.cgi?id=303460
Reviewed by Chris Dumez.

We were previously protecting the stream readers of GC for native sources in case the stream is pulling.
This does not always work with readers for streams teed from a native source.
We have to propagate the signal that the stream has a native source for teed streams, which we do in byteStreamTee.

We also have to extend the isReachableFromOpaqueRoots to a stream that is in readable state.
This is fin as we are only protecting readers with pending read requests, which means that
either pulling will happen in a very near future, or the pending read requests will be served.

Covered by updated LayoutTests/streams/blob-and-tee.html.

Canonical link: https://commits.webkit.org/303833@main