[threaded-animations] WPT test `scroll-animations/css/animation-inactive-outside-range-test.html` fails with "Threaded Scroll-driven Animations" enabled https://bugs.webkit.org/show_bug.cgi?id=303407 rdar://165710235
Reviewed by Simon Fraser.

We need to clamp the scroll offset for scroll timelines such that the current time does not
account for rubber-banding.

Note that there is no test change in this patch since the flag is not yet enabled on bots. This
was caught in preparation of that running animation tests locally using
`--experimental-feature ThreadedScrollDrivenAnimationsEnabled=true`.

* Source/WebCore/page/scrolling/ScrollingTreeScrollingNode.cpp:
(WebCore::ScrollingTreeScrollingNode::clampedCurrentScrollOffset const):
* Source/WebCore/page/scrolling/ScrollingTreeScrollingNode.h:
* Source/WebKit/UIProcess/RemoteLayerTree/RemoteProgressBasedTimeline.cpp:
(WebKit::RemoteProgressBasedTimeline::updateCurrentTime):

Canonical link: https://commits.webkit.org/303861@main