[Site Isolation] Fix WebPageProxy::broadcastFrameTreeSyncData to check that message comes from process containing frame rdar://165728342 https://bugs.webkit.org/show_bug.cgi?id=303437
Reviewed by Matt Woodrow.

WebPageProxy::{broadcastFrameTreeSyncData,broadcastAllFrameTreeSyncData}
contains a MESSAGE_CHECK that supposedly checks that the message is coming
from the process that owns the frame, according to the comment. But the
MESSAGE_CHECK itself actually checks if the process sending the message
is the main process. Fix the check to follow the comment.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::broadcastFrameTreeSyncData):
(WebKit::WebPageProxy::broadcastAllFrameTreeSyncData):

Canonical link: https://commits.webkit.org/303863@main