[GFC][Integration] Refactor positioning based eligibility https://bugs.webkit.org/show_bug.cgi?id=303449 rdar://165738264
Reviewed by Brent Fulgham.

Currently, we check the column and positioning style independently when
determining if we can take GFC. This is fine, but it seems like we will
want to check the start and end values together in certain scenarios as
we continue to relax these restrictions. For example, if we have an auto
column-start, we may only want to allow this if the column-end is an
explicit position or auto.

This patch refactors this code to better handle this dependency between
the start and end positioning properties. This is done by:

1.) Switching on the start (e.g. grid-column-start) property
2.) If we have a valid start, then calling a helper function to check
the corresponding end property based upon the type of start.

For example, now if we have an explicit grid-column-start and we meet
all other criteria, we will check the column-end by calling
hasValidColumnEnd. Currently, we have a hasValidColumnEnd/hasValidRowEnd
that takes in a Style::GridPositionExplicit, but as we continue to allow
more types of content, we can create more overloads of these functions to
check the column/row end corresponding to a specific type of start.

Canonical link: https://commits.webkit.org/303864@main