Add proper error forwarding for ReadableStream async iterable rdar://165607857 https://bugs.webkit.org/show_bug.cgi?id=303306
Reviewed by Chris Dumez.

We finalize the support of end of iteration by having ReadableStream::Iterator implement a isFinished getter.
This is usd by JSDOMAsyncIteratorBase to identify when the next promise resolves whether the iteration is done.
We then mark the IsFinish object as finished according ReadableStream::Iterator isFinished getter.
This is done in getNextIterationResult.

We also add a way to iterate asynchronously via DOMPromise:
- ReadableStream::Iterator::next returns a Ref<DOMPromise>
- JSDOMAsyncIteratorBase::getNextIterationResult will return the promise.
This allows ReadableStream::Iterator to forward any error that may be given to it.

We use IsAsyncIteratorNextReturningPromise concept to detect whether ReadableStream::Iterator::next is using the new code path.
We keep the old code path and plan to rework it in a follow-up.

All ReadableStream async iterator tests are now passing.

Covered by existing tests.

Canonical link: https://commits.webkit.org/303865@main